rename to set_algebra
